requirements.txt
PyGithub
requests
python-dotenv
torch
transformers
datasets
accelerate